back_end/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-phonenumber-field==7.1.0
psycopg==3.1.10
psycopg-binary==3.1.10
sqlparse==0.4.4
typing_extensions==4.7.1
